requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
openai==1.12.0
streamlit==1.32.0